# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Jewellery Shop API",
        "description": "Jewellery Shop Management API",
        "termsOfService": "https://jewellery-shop.test/terms",
        "contact": {
            "email": "admin@jewellery-shop.test"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API Server"
        },
        {
            "url": "/api/v1"
        }
    ],
    "paths": {
...
        "/reports/{report}/export": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Export report",
                "description": "Download a report as an Excel file",
                "operationId": "35529c73c09c974fde574cb76c4beee9",
                "parameters": [
                    {
                        "name": "report",
                        "in": "path",
                        "description": "Report type",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "sales",
                                "stock",
                                "pawn-outstanding",
                                "overdue-pawns",
                                "interest-earned",
                                "customer-ledger",
                                "supplier-ledger",
                                "profit"
                            ]
                        }
                    },
                    {
                        "name": "date_from",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "date_to",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "group_by",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "day",
                                "week",
                                "month"
                            ]
                        }
                    },
                    {
                        "name": "customer",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "supplier",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Excel file download",
                        "content": {
                            "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Unknown report"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}